refactoring tab code
