+++ src/modules/contact.js
@@ -1,0 +1,12 @@
+export default function infoContact (document) {
+    let contact = document.createElement('div');
+    let title = document.createElement('h3')
+    let text = document.createElement('p')
+
+    title.innerHTML = 'Contact'
+    text.innerHTML = 'sdihasiufhapiusdfhaisudbfipausdbf asdifubasdijf sd fasoid faishd fiahs dfh asdouf asoudf sid fajsh dfhjas dfouh asdof asidf asjid fuad udouas duof asdihf asuodhf uasdgfuasgdfioasdif asd fuasd fioasdbfisbdfouasbdfilas dfa so8dfbasidjf asd fasodfbsdfjbasidjfbaisdjbfiasdf asd9ufasdf asodif asi dfasudf iasudfijsdfu9is dfioausdbfas df8oasdf aisdfbas'
+
+    contact.appendChild(title)
+    contact.appendChild(text)
+    return contact
+}
