+++ src/modules/tabListener.js
@@ -1,27 +1,16 @@
 import infoMenu from './menu'
 import infoHome from './home'
+import infoContact from './contact'
+import clefill from './clean_fill'
+
 const tabListener =( (document) => {
 
-     const clear = (document) => {
-         
-         const info = document.getElementById('dividi')
-        info.innerHTML =''
-       
-        
-      };
 
-    const fill = (document, id, info) => {
-        
-         clear (document, id)
-        const fillInfo = document.getElementById('dividi')
-        fillInfo.appendChild(info)
-       
-    }
-    
-        
     const tabListener = (document, content) => {
-            
-            console.log(document)
+        let info =infoHome(document)
+         clefill.fill(document, info)
+         
+         
             const home = document.getElementById('home')
              const menu = document.getElementById('menu')
              const contact = document.getElementById('contact')
@@ -31,23 +20,18 @@ const tabListener =( (document) => {
                
                 do {
                     if (this_target === contact) {
-                        let info =infoContact(document)
-                    fill(document, contact, info)
-                    return;
+                         info =infoContact(document)
+                         
                     }else if (this_target === menu) {
-                        let info = infoMenu(document)
-                    fill(document, menu, info)
-                        return ;
+                         info = infoMenu(document)
                     } else if (this_target === home) {
-                        let info =infoHome(document)
-                        fill(document, home, info)
-                        return;
+                         info =infoHome(document)
+                        
                     }
-                
-                this_target = this_target.parentNode;
-               
+                    clefill.fill(document, info)
+                    return;
             } while (this_target);
-            clear(document,content)
+            clefill.clear(document,content)
             
                      
              });
@@ -59,21 +43,6 @@ const tabListener =( (document) => {
     
 })();
 
-
-
-
-const infoContact = (document) => {
-    let contact = document.createElement('div');
-    let title = document.createElement('h3')
-    let text = document.createElement('p')
-
-    title.innerHTML = 'Contact'
-    text.innerHTML = 'sdihasiufhapiusdfhaisudbfipausdbf asdifubasdijf sd fasoid faishd fiahs dfh asdouf asoudf sid fajsh dfhjas dfouh asdof asidf asjid fuad udouas duof asdihf asuodhf uasdgfuasgdfioasdif asd fuasd fioasdbfisbdfouasbdfilas dfa so8dfbasidjf asd fasodfbsdfjbasidjfbaisdjbfiasdf asd9ufasdf asodif asi dfasudf iasudfijsdfu9is dfioausdbfas df8oasdf aisdfbas'
-
-    contact.appendChild(title)
-    contact.appendChild(text)
-    return contact
-}
 
 
 export default tabListener
