+++ src/style.css
@@ -1,6 +1,6 @@
 #content {
-    color: red;
-   /* background: url('./restaurant.jpg'); */
+
+    background-color: blue;
    height: 100vh;
    text-align: center;
 }
